requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
stripe==7.11.0
twilio==8.12.0
gunicorn==21.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Werkzeug==3.0.1
email-validator==2.1.0.post1 